requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.4
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
kaitaistruct==0.10
numpy==2.1.1
outcome==1.3.0.post0
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydivert==2.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.1
requests==2.32.3
rsa==4.9
selenium==4.25.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
